requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.5
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0


